shape_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
_sysroot_linux-64_curr_repodata_hack=3=h69a702a_14
absl-py=2.1.0=pyhd8ed1ab_0
binutils_impl_linux-64=2.40=ha885e6a_0
binutils_linux-64=2.40=hdade7a5_3
brotli=1.1.0=hd590300_1
brotli-bin=1.1.0=hd590300_1
bzip2=1.0.8=hd590300_5
c-ares=1.28.1=hd590300_0
ca-certificates=2024.2.2=hbcca054_0
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.16.0=py312hf06ca03_0
chex=0.1.86=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.2.1=py312h8572e83_0
cuda-cccl_linux-64=12.5.39=ha770c72_0
cuda-crt-dev_linux-64=12.5.40=ha770c72_0
cuda-crt-tools=12.5.40=ha770c72_0
cuda-cudart=12.5.39=he02047a_0
cuda-cudart-dev=12.5.39=he02047a_0
cuda-cudart-dev_linux-64=12.5.39=h85509e4_0
cuda-cudart-static=12.5.39=he02047a_0
cuda-cudart-static_linux-64=12.5.39=h85509e4_0
cuda-cudart_linux-64=12.5.39=h85509e4_0
cuda-cupti=12.5.39=he02047a_0
cuda-driver-dev_linux-64=12.5.39=h85509e4_0
cuda-nvcc=12.5.40=hcdd1206_0
cuda-nvcc-dev_linux-64=12.5.40=ha770c72_0
cuda-nvcc-impl=12.5.40=hd3aeb46_0
cuda-nvcc-tools=12.5.40=hd3aeb46_0
cuda-nvcc_linux-64=12.5.40=h8a487aa_0
cuda-nvrtc=12.5.40=he02047a_0
cuda-nvtx=12.5.39=he02047a_0
cuda-nvvm-dev_linux-64=12.5.40=ha770c72_0
cuda-nvvm-impl=12.5.40=h59595ed_0
cuda-nvvm-tools=12.5.40=h59595ed_0
cuda-version=12.5=hd4f0392_3
cudnn=8.9.7.29=h092f7fd_3
cycler=0.12.1=pyhd8ed1ab_0
cython=3.0.10=py312h30efb56_0
fdasrsf=2.5.10=py312hfb8ada1_0
fonttools=4.51.0=py312h98912ed_0
freetype=2.12.1=h267a509_2
gcc_impl_linux-64=12.3.0=h58ffeeb_7
gcc_linux-64=12.3.0=h6477408_3
gxx_impl_linux-64=12.3.0=h2a574ab_7
gxx_linux-64=12.3.0=h4a1b8e8_3
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
jax=0.4.27=pyhd8ed1ab_0
jaxlib=0.4.23=cuda120py312h6027bbc_202
joblib=1.4.2=pyhd8ed1ab_0
kernel-headers_linux-64=3.10.0=h4a8ded7_14
kiwisolver=1.4.5=py312h8572e83_1
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.40=h55db66e_0
lerc=4.0.0=h27087fc_0
libabseil=20240116.2=cxx17_h59595ed_0
libblas=3.9.0=22_linux64_openblas
libbrotlicommon=1.1.0=hd590300_1
libbrotlidec=1.1.0=hd590300_1
libbrotlienc=1.1.0=hd590300_1
libcblas=3.9.0=22_linux64_openblas
libcublas=12.5.2.13=he02047a_0
libcufft=11.2.3.18=he02047a_0
libcurand=10.3.6.39=he02047a_0
libcusolver=11.6.2.40=he02047a_0
libcusparse=12.4.1.24=he02047a_0
libdeflate=1.20=hd590300_0
libexpat=2.6.2=h59595ed_0
libffi=3.4.2=h7f98852_5
libgcc-devel_linux-64=12.3.0=h0223996_107
libgcc-ng=13.2.0=h77fa898_7
libgfortran-ng=13.2.0=h69a702a_7
libgfortran5=13.2.0=hca663fb_7
libgomp=13.2.0=h77fa898_7
libgrpc=1.62.2=h15f2491_0
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=22_linux64_openblas
libllvm14=14.0.6=hcd5def8_4
libnsl=2.0.1=hd590300_0
libnvjitlink=12.5.40=he02047a_0
libopenblas=0.3.27=pthreads_h413a1c8_0
libpng=1.6.43=h2797004_0
libprotobuf=4.25.3=h08a7969_0
libre2-11=2023.09.01=h5a48ba9_2
libsanitizer=12.3.0=hb8811af_7
libsqlite=3.45.3=h2797004_0
libstdcxx-devel_linux-64=12.3.0=h0223996_107
libstdcxx-ng=13.2.0=hc0a3c3a_7
libtiff=4.6.0=h1dd3fc0_3
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.15=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=hd590300_5
llvmlite=0.42.0=py312hb06c811_1
matplotlib-base=3.8.4=py312h20ab3a6_2
ml_dtypes=0.4.0=py312h1d6d2e6_1
munkres=1.1.4=pyh9f0ad1d_0
nccl=2.21.5.1=h3a97aeb_0
ncurses=6.5=h59595ed_0
numba=0.59.1=py312hacefee8_0
numpy=1.26.4=py312heda63a1_0
openjpeg=2.5.2=h488ebb8_0
openssl=3.3.0=h4ab18f5_3
opt-einsum=3.3.0=hd8ed1ab_2
opt_einsum=3.3.0=pyhc1e730c_2
optax=0.2.2=pyhd8ed1ab_0
packaging=24.0=pyhd8ed1ab_0
pandas=2.2.2=py312h1d6d2e6_1
patsy=0.5.6=pyhd8ed1ab_0
pillow=10.3.0=py312hdcec9eb_0
pip=24.0=pyhd8ed1ab_0
pthread-stubs=0.4=h36c2ea0_1001
pycparser=2.22=pyhd8ed1ab_0
pyparsing=3.1.2=pyhd8ed1ab_0
python=3.12.3=hab00c5b_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.12=4_cp312
pytz=2024.1=pyhd8ed1ab_0
re2=2023.09.01=h7f4b329_2
readline=8.2=h8228510_1
scikit-learn=1.5.0=py312h1fcc3ea_1
scipy=1.13.1=py312hc2bc53b_0
setuptools=70.0.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sysroot_linux-64=2.17=h4a8ded7_14
threadpoolctl=3.5.0=pyhc1e730c_0
tk=8.6.13=noxft_h4845f30_101
toolz=0.12.1=pyhd8ed1ab_0
tqdm=4.66.4=pyhd8ed1ab_0
typing-extensions=4.11.0=hd8ed1ab_0
typing_extensions=4.11.0=pyha770c72_0
tzdata=2024a=h0c530f3_0
wheel=0.43.0=pyhd8ed1ab_1
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
zipp=3.17.0=pyhd8ed1ab_0
zstd=1.5.6=ha6fb4c9_0
